Refactor to move saveload button group into it's own component to make app more legible
